Address ServiceRequest comments
